feat(storybook): upgrade to storybook 6 - front-1483
