+++ src/systems/ec/implementations/react/storybook/.storybook/preview.js
@@ -1,4 +1,3 @@
-import { addDecorator, addParameters } from '@storybook/react';
 import { withA11y } from '@storybook/addon-a11y';
 import { INITIAL_VIEWPORTS } from '@storybook/addon-viewport';
 import { withCssResources } from '@storybook/addon-cssresources';
@@ -6,11 +5,7 @@ import { withCode } from '../../../../../../tools/storybook-addon-code';
 
 import './ECL';
 
-addDecorator(withCode);
-addDecorator(withA11y);
-addDecorator(withCssResources);
-
-addParameters({
+export const parameters = {
   a11y: {
     configure: {},
     options: {
@@ -86,4 +81,6 @@ html {
   options: {
     showRoots: true,
   },
-});
+};
+
+export const decorators = [withCode, withA11y, withCssResources];
