+++ src/systems/ec/implementations/react/storybook/package.json
@@ -16,13 +16,13 @@
     "@babel/plugin-proposal-export-namespace-from": "7.10.4",
     "@babel/runtime": "7.11.2",
     "@ecl/story-wrapper": "^2.34.0",
-    "@storybook/addon-a11y": "5.3.19",
-    "@storybook/addon-cssresources": "5.3.19",
+    "@storybook/addon-a11y": "6.0.26",
+    "@storybook/addon-cssresources": "6.0.26",
     "@storybook/addon-knobs": "5.3.19",
-    "@storybook/addon-viewport": "5.3.19",
-    "@storybook/addons": "5.3.19",
-    "@storybook/react": "5.3.19",
-    "@storybook/theming": "5.3.19",
+    "@storybook/addon-viewport": "6.0.26",
+    "@storybook/addons": "6.0.26",
+    "@storybook/react": "6.0.26",
+    "@storybook/theming": "6.0.26",
     "babel-core": "^7.0.0-bridge.0",
     "babel-loader": "8.1.0",
     "core-js": "3.6.5",
