+++ src/systems/eu/implementations/react/storybook/.storybook/preview.js
@@ -1,4 +1,3 @@
-import { addDecorator, addParameters } from '@storybook/react';
 import { withA11y } from '@storybook/addon-a11y';
 import { INITIAL_VIEWPORTS } from '@storybook/addon-viewport';
 import { withCssResources } from '@storybook/addon-cssresources';
@@ -6,11 +5,7 @@ import { withCode } from '../../../../../../tools/storybook-addon-code';
 
 import './ECL';
 
-addDecorator(withCode);
-addDecorator(withA11y);
-addDecorator(withCssResources);
-
-addParameters({
+export const parameters = {
   a11y: {
     configure: {},
     options: {
@@ -21,12 +16,12 @@ addParameters({
   cssresources: [
     {
       id: 'ecl-screen',
-      code: `<link rel="stylesheet" type="text/css" href="./styles/ecl-eu-preset-website.css" />`,
+      code: `<link rel="stylesheet" type="text/css" href="./styles/ecl-ec-preset-website.css" />`,
       picked: true,
     },
     {
       id: 'ecl-print',
-      code: `<link rel="stylesheet" type="text/css" href="./styles/ecl-eu-preset-website-print.css" />`,
+      code: `<link rel="stylesheet" type="text/css" href="./styles/ecl-ec-preset-website-print.css" />`,
       picked: false,
     },
     {
@@ -86,4 +81,6 @@ html {
   options: {
     showRoots: true,
   },
-});
+};
+
+export const decorators = [withCode, withA11y, withCssResources];
